+++ CHANGES.rst
@@ -23,6 +23,12 @@ Release Notes
 
 - Fixed issue where a non-string assigned to window.id caused
   'Select Window' and 'Get Window *' keywords to fail.
+  [ombre42]
+
+- Allow using key attributes (default strategy) when the locator contains
+  a '=' by using the prefix 'default='. Also make locator prefixes
+  space-insensitive.
+  [ombre42]
 
 A big thank you to [eweitz] and [HelioGuilherme66] for getting the
 continuous integration builds to go green by fixing internal tests.